Moved init related tasks and plugins to appropriate package
