+++ src/main/groovy/org/shipkit/internal/gradle/ReleaseConfigurationPlugin.java
@@ -4,8 +4,9 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.ObjectConfigurationAction;
-
 import org.shipkit.gradle.ReleaseConfiguration;
+import org.shipkit.internal.gradle.init.InitConfigFileTask;
+import org.shipkit.internal.gradle.init.InitPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
 import org.shipkit.internal.version.VersionInfo;
