+++ src/main/groovy/org/shipkit/internal/gradle/VersioningPlugin.java
@@ -7,6 +7,8 @@ import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.shipkit.gradle.BumpVersionFileTask;
 import org.shipkit.internal.gradle.git.GitPlugin;
+import org.shipkit.internal.gradle.init.InitPlugin;
+import org.shipkit.internal.gradle.init.InitVersioningTask;
 import org.shipkit.internal.gradle.util.TaskMaker;
 import org.shipkit.internal.version.Version;
 import org.shipkit.internal.version.VersionInfo;
