+++ src/main/groovy/org/shipkit/internal/gradle/init/InitPlugin.java
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle;
+package org.shipkit.internal.gradle.init;
 
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
