+++ src/test/groovy/org/shipkit/internal/gradle/VersioningPluginTest.groovy
@@ -3,6 +3,7 @@ package org.shipkit.internal.gradle
 import org.shipkit.gradle.BumpVersionFileTask
 import org.shipkit.internal.gradle.git.GitCommitTask
 import org.shipkit.internal.gradle.git.GitPlugin
+import org.shipkit.internal.gradle.init.InitVersioningTask
 import org.shipkit.internal.version.VersionInfo
 import testutil.PluginSpecification
